fix(storage): guard binary payloads by the buffer protocol, not a type list
A `memoryview` reached `canonical_form`'s `Sequence` branch and encoded as
`[114, 97, 119]`, taking the same semantic digest as a genuine integer list
— an inline binary payload silently reinterpreted, and deduplicated against
unrelated logical content in a content-addressed store. Reproduced through
both public entry points, since each routes through `canonical_form`.

The guard listed `bytes` and `bytearray`, so it was only as complete as the
list. `array.array` and `mmap.mmap` are the same shape, and the next such
type would have been the third instance of one bug — so the test is now
whether a value exposes a buffer at all, which is the question the rule is
actually about. The message names the type passed rather than always
"bytes", so a caller is not sent looking for one they never wrote.

Splitting the tests: `test_canonical.py` crossed the 1200-line test cap, so
D2's version axes moved to `test_versioning.py`. The line is the one that
file's own docstring had already drawn by announcing two subjects, and the
test count is unchanged across the move.
